+++ tools_fix_bloated_merged_blocks.go
@@ -3,26 +3,24 @@ package firecore
 import (
 	"fmt"
 	"io"
-	"strconv"
 
 	"github.com/spf13/cobra"
 	"github.com/streamingfast/bstream"
-	"github.com/streamingfast/cli"
 	"github.com/streamingfast/dstore"
 	"github.com/streamingfast/firehose-core/tools"
 	"go.uber.org/zap"
 )
 
-func newToolsUnmergeBlocksCmd[B Block](chain *Chain[B], zlog *zap.Logger) *cobra.Command {
+func newToolsFixBloatedMergedBlocks[B Block](chain *Chain[B], zlog *zap.Logger) *cobra.Command {
 	return &cobra.Command{
-		Use:   "unmerge-blocks <src_merged_blocks_store> <dest_one_blocks_store> [<block_range>]",
-		Short: "Unmerges merged block files into one-block-files",
+		Use:   "fix-bloated-merged-blocks <src_merged_blocks_store> <dest_one_blocks_store> [<block_range>]",
+		Short: "Fixes 'corrupted' merged-blocks that contain extraneous or duplicate blocks. Some older versions of the merger may have produce such bloated merged-blocks. All merged-blocks files in given range will be rewritten, regardless of if they were corrupted.",
 		Args:  cobra.ExactArgs(3),
-		RunE:  runUnmergeBlocksE(zlog),
+		RunE:  runFixBloatedMergedBlocksE(zlog),
 	}
 }
 
-func runUnmergeBlocksE(zlog *zap.Logger) CommandExecutor {
+func runFixBloatedMergedBlocksE(zlog *zap.Logger) CommandExecutor {
 	return func(cmd *cobra.Command, args []string) error {
 		ctx := cmd.Context()
 
@@ -67,6 +65,18 @@ func runUnmergeBlocksE(zlog *zap.Logger) CommandExecutor {
 				return fmt.Errorf("creating block reader: %w", err)
 			}
 
+			mergeWriter := &mergedBlocksWriter{
+				store:         destStore,
+				writerFactory: bstream.GetBlockWriterFactory,
+				tweakBlock:    func(b *bstream.Block) (*bstream.Block, error) { return b, nil },
+				logger:        zlog,
+			}
+
+			seen := make(map[string]bool)
+
+			var lastBlockID string
+			var lastBlockNum uint64
+
 			// iterate through the blocks in the file
 			for {
 				block, err := br.Read()
@@ -74,7 +84,7 @@ func runUnmergeBlocksE(zlog *zap.Logger) CommandExecutor {
 					break
 				}
 
-				if block.Number < uint64(blockRange.Start) {
+				if block.Number < uint64(startBlock) {
 					continue
 				}
 
@@ -82,39 +92,23 @@ func runUnmergeBlocksE(zlog *zap.Logger) CommandExecutor {
 					break
 				}
 
-				oneblockFilename := bstream.BlockFileNameWithSuffix(block, "extracted")
-				zlog.Debug("writing block", zap.Uint64("block_num", block.Number), zap.String("filename", oneblockFilename))
-
-				pr, pw := io.Pipe()
-
-				//write block data to pipe, and then close to signal end of data
-				go func(block *bstream.Block) {
-					var err error
-					defer func() {
-						pw.CloseWithError(err)
-					}()
-
-					var bw bstream.BlockWriter
-					bw, err = bstream.GetBlockWriterFactory.New(pw)
-					if err != nil {
-						zlog.Error("creating block writer", zap.Error(err))
-						return
-					}
-
-					err = bw.Write(block)
-					if err != nil {
-						zlog.Error("writing block", zap.Error(err))
-						return
-					}
-				}(block)
-
-				//read block data from pipe and write block data to dest store
-				err = destStore.WriteObject(ctx, oneblockFilename, pr)
-				if err != nil {
-					return fmt.Errorf("writing block %d to %s: %w", block.Number, oneblockFilename, err)
+				if seen[block.Id] {
+					zlog.Info("skipping seen block (source merged-blocks had duplicates, skipping)", zap.String("id", block.Id), zap.Uint64("num", block.Number))
+					continue
 				}
 
-				zlog.Info("wrote block", zap.Uint64("block_num", block.Number), zap.String("filename", oneblockFilename))
+				if lastBlockID != "" && block.PreviousId != lastBlockID {
+					return fmt.Errorf("got an invalid sequence of blocks: block %q has previousId %s, previous block %d had ID %q, this endpoint is serving blocks out of order", block.String(), block.PreviousId, lastBlockNum, lastBlockID)
+				}
+				lastBlockID = block.Id
+				lastBlockNum = block.Number
+
+				seen[block.Id] = true
+
+				fmt.Println("processing", block.String())
+				if err := mergeWriter.ProcessBlock(block, nil); err != nil {
+					return fmt.Errorf("write to blockwriter: %w", err)
+				}
 			}
 
 			return nil
@@ -130,11 +124,4 @@ func runUnmergeBlocksE(zlog *zap.Logger) CommandExecutor {
 
 		return nil
 	}
-}
-
-func mustParseUint64(s string) uint64 {
-	i, err := strconv.Atoi(s)
-	cli.NoError(err, "Unable to parse %q as uint64", s)
-
-	return uint64(i)
 }